bugfix: SplitCiphertextTag did not handle tag-only files.
The SplitCiphertextTag function had an off-by-one error where it
incorrecly expected the ciphertext to have at least one non-tag
byte.  This is incorrect, as encrypting a 0-byte file (or, more
generally, an empty buffer of data) results in an output
that is only the 16-byte integrity tag.  The fix now supports
such a scenario where the "ciphertext" only consists of the tag.